Restructure DB design, Changes all the required entities and services

+++ src/main/java/in/lekhai/core/dto/admin/AdminRegistrationResponse.java
@@ -9,5 +9,5 @@ public record AdminRegistrationResponse(
         @JsonIgnore
         String uuid,
         @JsonIgnore
-        Long categoryId
+        Integer categoryId
 ) { }
