+++ src/main/java/in/lekhai/core/dto/category/EnableCategoryWiseFeatures.java
@@ -3,7 +3,7 @@ package in.lekhai.core.dto.category;
 import java.util.Set;
 
 public record EnableCategoryWiseFeatures (
-        Set<Long> categoryIdList,
+        Set<Integer> categoryIdList,
         Set<Integer> bitsPositionsToBeEnabled
 )
 { }
